removing unused make targets and legacy code
