+++ Makefile
@@ -155,22 +155,8 @@ docker-e2e-test: e2e ## Run docker integration test in new ec2 instance
 e2e-cleanup: e2e ## Clean up resources generated by e2e tests
 	scripts/e2e_cleanup.sh
 
-.PHONY: vsphere-e2e-test
-vsphere-e2e-test: build-cross-platform ## Run vsphere integration tests
-	go run cmd/integration_test/main.go e2e vsphere \
-		-a ${INTEGRATION_TEST_UBUNTU_AMI_ID} \
-		-s ${INTEGRATION_TEST_STORAGE_BUCKET} \
-		-j ${JOB_ID} \
-		-i ${INTEGRATION_TEST_INSTANCE_PROFILE} \
-		-n ${INTEGRATION_TEST_SUBNET_ID} \
-		-d false
-
-.PHONY: build-e2e-test-binaries
-build-e2e-test-binaries: build build-cross-platform ## Build e2e binaries
-	go build -o bin/eks-a-e2e github.com/aws/eks-anywhere/cmd/integration_test
-
 .PHONY: capd-test-all
-capd-test-all: capd-test-118 capd-test capd-test-120
+capd-test-all: capd-test capd-test-120
 
 .PHONY: capd-test-%
 capd-test-%: e2e ## Run CAPD tests
@@ -299,14 +285,6 @@ $(RELEASE_DIR):
 release-manifests: $(KUSTOMIZE) generate-manifests $(RELEASE_DIR) ## Builds the manifests to publish with a release
 	# Build core-components.
 	$(KUSTOMIZE) build config/prod > $(RELEASE_DIR)/eksa-components.yaml
-
-.PHONY: e2e-cleanup-postsubmit # Temporary target to make postsubmit cleanup job work.
-e2e-cleanup-postsubmit:
-	go run cmd/integration_test/main.go e2e cleanup \
-		-s ${INTEGRATION_TEST_STORAGE_BUCKET} \
-		-a ${INTEGRATION_TEST_MAX_INSTANCE_AGE} \
-		-t ${INTEGRATION_TEST_INSTANCE_TAG} \
-		-d false
 
 .PHONY: run-controller # Run eksa controller from local repo with tilt
 run-controller:
